+++ packages/health/example/lib/main.dart
@@ -31,13 +31,13 @@ class _MyAppState extends State<MyApp> {
   }
 
   Future fetchData([bool mayUseLastAccount = true]) async {
-    /// Get everything from midnight until now
+    /// get everything from midnight until now
     DateTime startDate = DateTime(2020, 11, 07, 0, 0, 0);
     DateTime endDate = DateTime(2025, 11, 07, 23, 59, 59);
 
     HealthFactory health = HealthFactory();
 
-    /// Define the types to get.
+    // define the types to get
     List<HealthDataType> types = [
       HealthDataType.STEPS,
       HealthDataType.WEIGHT,
@@ -65,21 +65,21 @@ class _MyAppState extends State<MyApp> {
       prefs.setString(_accountNamePreference, accountName);
 
       try {
-        /// Fetch new data
+        // fetch new data
         List<HealthDataPoint> healthData =
           await health.getHealthDataFromTypes(startDate, endDate, types, accountName);
 
-        /// Save all the new data points
+        /// save all the new data points
         _healthDataList.clear();
         _healthDataList.addAll(healthData);
       } catch (e) {
         print("Caught exception in getHealthDataFromTypes: $e");
       }
 
-      /// Filter out duplicates
+      // filter out duplicates
       _healthDataList = HealthFactory.removeDuplicates(_healthDataList);
 
-      /// Print the results
+      // print the results
       _healthDataList.forEach((x) {
         print("Data point: $x");
         steps += x.value.round();
@@ -87,7 +87,7 @@ class _MyAppState extends State<MyApp> {
 
       print("Steps: $steps");
 
-      /// Update the UI to display the results
+      // update the UI to display the results
       setState(() {
         _state =
             _healthDataList.isEmpty ? AppState.NO_DATA : AppState.DATA_READY;