+++ packages/light/example/lib/main.dart
@@ -14,7 +14,7 @@ class _MyAppState extends State<MyApp> {
   late Light _light;
   late StreamSubscription _subscription;
 
-  void onData(int? luxValue) async {
+  void onData(int luxValue) async {
     print("Lux value: $luxValue");
     setState(() {
       _luxString = "$luxValue";
@@ -28,9 +28,8 @@ class _MyAppState extends State<MyApp> {
   void startListening() {
     _light = new Light();
     try {
-      _subscription = _light.lightSensorStream!.listen(onData);
-    }
-    on LightException catch (exception) {
+      _subscription = _light.lightSensorStream.listen(onData);
+    } on LightException catch (exception) {
       print(exception);
     }
   }
@@ -51,10 +50,10 @@ class _MyAppState extends State<MyApp> {
     return new MaterialApp(
       home: new Scaffold(
         appBar: new AppBar(
-          title: const Text('Plugin example app'),
+          title: const Text('Light Example App'),
         ),
         body: new Center(
-          child: new Text('Running on: $_luxString\n'),
+          child: new Text('Lux value: $_luxString\n'),
         ),
       ),
     );