+++ packages/activity_recognition_flutter/CHANGELOG.md
@@ -1,3 +1,9 @@
+## 4.0.4
+* improvements to documentation
+
+## 4.0.3
+* [PR #358](https://github.com/cph-cachet/flutter-plugins/pull/358)
+
 ## 4.0.2
 * [PR #302](https://github.com/cph-cachet/flutter-plugins/pull/302)
 * [PR #351](https://github.com/cph-cachet/flutter-plugins/pull/351)