+++ packages/light/CHANGELOG.md
@@ -1,11 +1,12 @@
-## 2.0.0
+## 2.1.0
+- `lightSensorStream` is not returning `null` values.
+- update to example app
 
+## 2.0.0
 - Migration to null safety
 
 ## 1.0.1
-
 - fix: issue [#139](https://github.com/cph-cachet/carp.sensing-flutter/issues/139)
 
 ## 1.0.0
-
 - Updated version of the old plugin.